requirements.txt
discord.py
python-dotenv
pytz
requests
ffmpeg
ffmpeg-python
PyNaCl
pip
youtube-dl
discord.py[voice]
discord-ext-audiorec
stripe
paypalrestsdk